Refactor_2024.12.07.00.52_MatrixColコンポーネントのリファクタリング
